Update org-roam faces
This matches the current state of the project.

The main colour of org-roam links is now different than standard links.
It is an attempt to differentiate between backlinks to other parts of
the knowledge base and whatever else a user may link to.  If this is not
desired, you can evaluate the following:

    (setq org-roam-link-use-custom-faces nil)