Simplify Org heading extra faces to match upstream
This was a planned change to align the themes with the status of
upstream Org.  For details, read my article on adaptable Org heading
faces: <https://protesilaos.com/codelog/2020-09-24-org-headings-adapt/>.
It concerns Org's "master" branch and will become widely available in
version 9.5.

It is known that those adjustments will break some of the customisation
styles for headings on older versions of Org (see manual for headings:
<https://protesilaos.com/modus-themes/#h:271eff19-97aa-4090-9415-a6463c2f9ae1>).
I am, however, prepared to accept this state of affairs because we
should not try to awkwardly cover bugs with ad-hoc workarounds.

This also fixes unintended side-effects, such as those presented in
issue 104: <https://gitlab.com/protesilaos/modus-themes/-/issues/104>.
Thanks to user "bepolymathe" for reporting the issue.