New custom option for comment colour
* modus-operandi-theme.el (modus-operandi-theme-comments): New defcustom
to control the colour of comments.
(modus-operandi-theme-comment): Helper function to use the correct
colour depending on the value of the new defcustom.
(custom-theme-set-faces): Implement new option and make it consistent
across various face groups.

* modus-operandi-theme.el (File): Same as above.

* doc/modus-themes.org (Option for comments): Document new customisation
option.
(Configure options prior to loading): Include new customisation option
in sample configuration.

* doc/modus-themes.info (File): Recompile manual.

This is in response to feedback I received in issue 105:
<https://gitlab.com/protesilaos/modus-themes/-/issues/105>.  Thanks to
Eugene and Manuel Uberti for their comments.