Ensure uniformity of various bold/italic faces
The idea here is that a user may wish to apply a different set of
attributes to everything that is "bold" and/or "italic".  For example,
to use a semibold weight for bold elements, or to apply a different
typeface to italic ones.

By ensuring consistency, the user can now evaluate the following after
loading the theme:

    ;; Example for semibold
    (set-face-attribute 'bold nil :weight 'semibold)

    ;; Example for distinct italic family
    (set-face-attribute 'italic nil :family "Source Code Pro")

Also see some major preparatory work in commits 4c4dccb50, 76b631f8c.