Disable byte compilation until further notice
We are facing a fundamental problem with the byte compilation of the
defmacro found in the source code.  That macro expands the palette alist
inside of a 'let' form and then binds it to all face definitions.  It
also passes another list, empty by default, which is meant for end users
who wish to override certain colours (or create new ones).

The reason we have a problem with byte compilation appears to be this:
the macro is expanded right after it is defined, at which point the
compiler throws an error that the two alists passed to the macro are
void as variables.  This only affects compiled code, because expanding
the defmacro once everything is loaded does produce the expected code,
namely: face definitions wrapped inside a 'let' call that locally binds
the contents of the alists (the colour associations).

What we need is to somehow inform the compiler that the alists exist at
the time the macro is expanded in compiled code.  But we also want to
make sure that users can still access the alist with the colours and can
still write their own overrides.

Disabling byte compilation is a compromise I must make for the themes to
continue to work as intended, albeit at a performance cost in loading
times at startup.

If you have knowledge of how to provide a genuine solution to this
problem, please help.

This commit should "solve" issue 102:
https://gitlab.com/protesilaos/modus-themes/-/issues/102.

It also relates to issue 109, though I do not know whether it also fixes
it: https://gitlab.com/protesilaos/modus-themes/-/issues/109.