Decouple faces from variables
This fixes the faulty macro expansion of nesting the call to
'custom-theme-set-variables' inside 'custom-theme-set-faces'.  Actually
you could not catch the error just by using the theme because it
worked as expected...

Also makes the code easier to understand, as it clearly separates faces
from variables (defface VS defvar or defcustom).