New custom option: modus-themes-paren-match
WARNING for early adopters of the yet-to-be-released version 1.0.0:[1]
the 'modus-themes-intense-paren-match' is no longer present and I did
not write a declaration to render it obsolete.  The idea is that all the
'make-obsolete' should be about the transition from the older versions
to 1.0.0.  Thank you for your understanding and apologies for the
inconvenience.

[1]: https://gitlab.com/protesilaos/modus-themes/-/issues/119