Add display spec to modus-themes-with-colors
This makes it possible to use face definitions that are written in the
exact same way as in the themes' source code.  For example

    `(default ((,class :background ,bg-main :foreground ,fg-main)))

Thanks to Gustavo Baros for bringing the issue to my attention:
<https://gitlab.com/protesilaos/modus-themes/-/issues/136#note_475240733>.

And thanks again to Nicolas De Jaeghere for contributing the
'modus-themes-with-colors' macro (commit 4ef50ceef).