Remove obsolete Consult faces; add a new one
There are some more Consult faces whose defaults are aligned with the
objective of the themes: they inherit from basic faces, such as 'error'
and 'success', which are already styled properly.  As such, there is no
pressing need to cover them outright.

Thanks to Daniel Mendler (Consult's maintainer) for reporting this in
issue in 155: <https://gitlab.com/protesilaos/modus-themes/-/issues/155>.