Refine all email citation faces
The intent is to tone down a bit the combined colouration of the lower
quote levels.  So the typical scenario of replying to an email and
citing one through four messages in the thread, will result in a
relatively less vibrant aesthetic than before.

All other colours are adapted accordingly.