MAJOR REVIEW of the manual's formatting for texi
The idea is to improve the output of the manual in Texinfo, so that it
is easier for me to send it to upstream Emacs with the proper markup.

While core Emacs should be covered, there still are some issues I need
to deal with, as those 'kbd' and 'file' macros do not export any markup
for the HTML backend.  Ideally, those should be interpreted as <code>,
though I do not really mind if that is not doable because I only publish
the HTML on my website where this lack of markup does not really affect
the meaning of the instructions.

This is done in response to Emacs bug#45143.