Add (setq modus-themes-diffs 'deuteranopia)
This partially addresses the wishlist entry to provide options that
account for color diffiency (thanks to Hörmetjan Yiltiz):
<https://gitlab.com/protesilaos/modus-themes/-/issues/118>.

I was also inspired to implement this feature after reading Sam
Kleinman's blog post on "Editor Themes" (2021-02-24):
<https://tychoish.com/post/editor-themes/>.

Apart from what this new option provides, please consider this a
statement of intent: to make the themes cope well with such
requirements.  There is more work to be done on that front and any user
feedback would be most welcome.