Refine modus-vivendi bg-{alt,dim} for better TUI
On terminal emulators with limited colour support, the old values would
be represented as shades of blue.  This was particularly problematic
when displaying inactive modelines or line numbers.

The new colour values should always be recalculated as the nearest grey.
The adjustment is marginal and should not have any impact on GUI Emacs.