Simplify manual's kbd macro for Emacs26 build
This has been prompted by an issue in the GNU ELPA build system which
cannot derive an Info manual from this Org file on Emacs26.  The
simplified format of the macro is enough for my case, it fixes the issue
with Emacs26, and has no effect whatsoever on the generated output (I
manually checked the exported .texi files with 'ediff-files').