Refine bongo faces for consistency
The previous design was meant to keep the track fields distinct from
each other.  However the use of yellow was not good aesthetically: it
does not fit with the rest of the theme.

Upon further experimentation, I realised that the album field (yellow)
is only present when the artist and title fields are also available:
first is the title, then the artist, and finally the album.  This is
true even with 'bongo-join-inserted-tracks' set to a non-nil value.  So
changing the face from yellow to a neutral value is safe.

The other two faces are adapted to look better in the new context.