Remove old deprecated options (prior to 1.0.0)
Users who could be affected by this will be using old GNU ELPA packages
up to version 0.12.0 (summer 2020) or earlier.  Since then we have
updated GNU ELPA to 1.2.0 and 1.3.0, so there is no need for the next
version (1.4.0) to keep this code.

I am pushing this early in the development cycle of 1.4.0 so that we get
plenty of time to deal with any possible breakage.