Make more diff faces/packages respect deuteranopia
This makes them conform with either of those:

     (setq modus-themes-diffs 'deuteranopia)
     (setq modus-themes-diffs 'fg-only-deuteranopia)

Making the themes work well for users with red-green colour deficiency
is a realisable (and laudable) objective that is consistent with our
commitment to uncompromising accessibility.

Given the broad package coverage, this remain work-in-progress that
consists of multiple commits.  The original idea was proposed in issue
118 by Hörmetjan Yiltiz:
<https://gitlab.com/protesilaos/modus-themes/-/issues/118>.