New defcustom modus-themes-success-deuteranopia
I think we now are in a position to claim with confidence that the Modus
themes are committed to empower users with red-green colour deficiency.

From now on, every context that has a red/green dichotomy of some sort
and which has no alternative of a red/blue binary should be considered a
serious bug against the themes.  I take this issue as seriously as the
primary objective of the themes for a WCAG AAA standard for legibility.

With this new option, as well as everything that has done into
'modus-themes-diffs' (e.g. the earlier commit 133d64e3a9 or the review
of its "fg only" option[1]), we can finally claim that issue 118 is
addressed, at least in principle.[2]

[1]: <https://gitlab.com/protesilaos/modus-themes/-/issues/183>.
[2]: <https://gitlab.com/protesilaos/modus-themes/-/issues/118>.