Add support for built-in pulse faces
The 'pulse-highlight-face' warns in its doc string not to customise it,
though searching through the commit logs in emacs.git as well as greping
the source does not yield any answer as to why that is.  The original
commit was in 2009 and it provides no justification whatsoever.

As such, we decide to customise that face regardless: the worst that
could happen is some subtle bug somewhere (i.e. a different default
colour than expected), which we are fully prepared to deal with.

Thanks to Gustavo Barros for reporting this issue and helping search
through the logs to decide on this course of action:
<https://gitlab.com/protesilaos/modus-themes/-/issues/185>.