Abstract search faces; respect deuteranopia option
The abstraction is done in an effort to ensure consistency between
different face groups, but also to make it easier to propagate
customisation options.

The one option that applies now is for deuteranopia.  It means to use a
blue colour instead of a green one to denote a successful state when
this is evaluated:

    (setq modus-themes-success-deuteranopia t)

For more on this new variable check commit 70456e93f5.