Tweak org-quote for modus-themes-org-blocks
The default is a subtle blue/cold foreground colour against the main
background (black/white).  When 'modus-themes-org-blocks' is set to a
'grayscale' value (meaning that the block has a gray background), we
change its foreground to the main foreground colour in order to maintain
a good level of legibility.

This is done based on the feedback I received from Rudolf Adamkovič in
issue 190: <https://gitlab.com/protesilaos/modus-themes/-/issues/190>.