Simplify mail citations
Gnus has 11(!) levels of depth for citations.  I think that (i) such
cases are ugly to read in general no matter what colours are used,
and (ii) four alternating colours are enough to tell apart one level
from the next.

We inherit from the message faces because all those packages have the
form "(require 'message)" so we are certain that the faces will be
available.