Change value of modus-themes-mail-citations option
The "faint" is more accurate than "desaturated".  To avoid breaking the
setup of current users, the latter will remain a valid option for the
foreseeable future (all such cases will be reviewed ahead of the
eventual version 2.0.0 of the themes).

Thanks to André Alexandre Gomes for the feedback in issue 193:
<https://gitlab.com/protesilaos/modus-themes/-/issues/193>.