Remove support for pulse.el until further notice
This is in response to issue 200, reported by Gustavo Barros:
<https://gitlab.com/protesilaos/modus-themes/-/issues/200>

In essence, we are not sure whether the themes may be causing
unpredictable behaviour with the ':extend' property of the pulse faces,
which sometimes stops working.

The doc string of pulse-highlight-face does not give us confidence:

    Face used during a pulse for display.  *DO NOT CUSTOMIZE*
    Face used for temporary highlighting of tags for effect.

However we could not find a case in emacs.git where the face is reset or
otherwise changed.  Furthermore, it is not clear how theme support for
it can cause a problem: the worst that could happen is for the
overriding attributes to be applied, though we have no indication of
that.