Subject Mu4e current line to modus-themes-hl-line
This is to ensure consistency across the theme, wherever an in-buffer
current line highlight is present.

The variable 'modus-themes-hl-line' can be configured to make the
current line highlight more intense, add an accent colour to it, and/or
optionally apply an underline as well.  Please consult the variable's
doc string.

The 'mu4e-header-highlight-face' will now use only a background colour
as it did before.  So it will not override any text highlights on the
current line.  This is important for usability, such as to differentiate
ordinary read mail headers from replied ones.

This face was mistakenly edited in commit a53f23fcc3, as a response to
issue 214: <https://gitlab.com/protesilaos/modus-themes/-/issues/214>.
Thanks to Gustavo Barros for further feedback which highlighted some of
the problems with the temporary change (and which reminded me that
making Mu4e inconsistent with the rest of the theme by rendering its
current line uncustomisable was the wrong course of action---mea culpa).