Don't apply diff context colour in bg-only style
This concerns the optional setting (setq modus-themes-diffs 'bg-only)
which refrains from overriding any foreground colour in the diff buffer.
It is meant to be used either as an aesthetic preference or as a
usability enhancement when 'diff-font-lock-syntax' is set to a non-nil
value (Magit does not have an equivalent yet, but we should expect it to
be implemented eventually).

Also check issue 218 in case you want to share your thoughts on this
style and on whether it should become the new default in some future
version of the themes: <https://gitlab.com/protesilaos/modus-themes/-/issues/218>.