Improve performance of modus-themes--set-option
When modus-theme user options are customized and added to the user
face, loading the theme would unintentionally reload the theme
multiple times, leading to an unnecessary performance drop.

This change checks if `enable-theme` is loading the current theme, and
inhibits the reloading feature, to avoid the slow-down, while
preserving the interactive reloading the improves the discoverability
of customization options for users.