Add custom face for the consistency of all prompts
Having the prompt faces inherit from 'comint-highlight-prompt' could
lead to undefined faces if the comint library were not loaded.  Whereas
the 'modus-themes-prompt' guarantees that we pass a known face at all
times.

Thanks to Philip Kaludercic for bringing this potential bug to my
attention in a comment to merge request 43:
<https://gitlab.com/protesilaos/modus-themes/-/merge_requests/43#note_615224855>.