Refine diredfl and dired+ faces
For the various "priv" faces, the intent is to increase the difference
in hueness between adjacent file permissions (the changes are minor, but
they do change the overall result).

Numbers are toned down so that they do not clash with dates.

The file suffix no longer uses cyan to stand out more in detailed views.