Revise marginalia faces (extend 66fbe78c5a)
In commit 66fbe78c5a we added support for some new Marginalia faces.
Those use the same styles as diredfl/dired+, from whence they draw
inspiration (see commit 72b3793fc4).

To ensure consistency in terms of colouration and overall look and feel
between the numerous Marginalia listings, we need to change all the
other faces as well.  The intent is two-fold:

1. Use colour combinations which stand out and which do not rely on
sharp contrast in hueness between adjacent entries.  We do not want to
make annotations compete for attention with completion candidates.  The
latter must be prioritised in the context of completion.

2. Make the various Marginalia views more akin to each other, so that
they feel part of the same experience.