Inherit Emacs28 key "button" if possible
In Emacs 28 all key bindings are automatically displayed with the face
help-key-binding which uses some new face attributes to draw a cleaner
box around it.  Given that the themes must work with earlier versions,
we cannot inherit it unconditionally.  It is nice to offer this feature
to those who use the themes on Emacs 28.