Align ediff with other modus-themes-diffs styles
The notion of a "focus" state only make sense in Magit which applies
colour-coding to all diff hunks.  Whereas non-active Ediff changes are
grayed out, so there is no need for more colour-coding nuances.

Those styles had stayed behind even after reviewing all Ediff faces in
commits f8df057, 98df906, 7572c60.  The most noticeable change is with:

    (setq modus-themes-diffs 'bg-only)

* * *

This commit overrides the change made in 29fd33c.  I think the addition
of a subtle neutral background is not required in that case.

If users think this change is for the worse, we can always define a
helper function like this:

    (defun modus-themes--ediff-style (bgonly default)
      "Diff style for Ediff.
    BGONLY and DEFAULT depend on the value of `modus-themes-diffs'.
    The former is more subtle."
      (if (eq modus-themes-diffs 'bg-only)
          (list bgonly)
        (list default)))

And apply it thus:

    `(ediff-current-diff-A ((,class :inherit ,@(modus-themes--ediff-style
                                                'modus-themes-diff-removed
                                                'modus-themes-diff-focus-removed))))