Note that modus-themes-lang-checkers is for 29.1
This is because the themes are built into Emacs.  Because the emacs-28
branch has already been cut, the next time I sync the themes with
emacs.git it will be on the 'master' branch, which is the development
target for Emacs29.

The modus-themes-lang-checkers was last changed in commit 32e86cd.