Use 'shadow' instead of subtle fg, where relevant
This makes it easier for users to customise the 'shadow' face to e.g. a
"gray50" foreground value and have the change apply across the theme.

There still are some cases where we retain the use of fg-alt, because
those involve a background or are otherwise context-dependent and should
not be changed as easily.