{Internal} Do not return nil mode-line attributes
Depending on the value of modus-themes-mode-line, the helper function
modus-themes--mode-line-attrs would return a nil value for some
attributes.  In most cases, these do not matter.  There is, however, a
chance that some code is unconditionally using something like:

    (face-attribute 'mode-line :box)

If the return value is nil, that will cause problem.  Whereas the
special 'unspecified' symbol has the meaning of an empty attribute and
is a non-nil value.