Bring back box style for transient keys
We had reverted that style in commit 83aaca4, as it caused misalignments
in the transient window.  But since commit bf132a5 this is not longer
the case, so we can safely have theme-wide consistency for all key
bindings.

Note that the box style uses negative margins to not affect the line's
width/height.  This only applies to users of Emacs 28 or higher.  As
such, for Emacs 27 we make the keys have a blue colour and be rendered
in a bold font weight.

Thanks to Manuel Uberti and Kevin Fleming for the feedback in issue 232:
<https://gitlab.com/protesilaos/modus-themes/-/issues/232>.