Make the org-colview faces always same height
In order for columns and headings to line up when using org-mode columns, even
when modus-themes-scale-headings is enabled, make both relevant faces inherit
from the default face, so that they use the default height.

It's done this way because if the org-column :height is set to a floating point
number, that's applied to the underlying heading size, not to the default
height.

In addition, making headings relatively small when in column view is sensible as
in that mode more information needs to fit on the same row.