BREAKING: Consolidate all deuteranopia styles
Everything should now be subject to a single toggle:
modus-themes-deuteranopia.  The modus-themes-success-deuteranopia is
thus superseded, while the individual deuteranopia-friendly styles for
diffs and the Org agenda's habit graph are altogether removed (there is
no clean way to notify the user of their deprecation---apologies for the
inconvenience!).

* * *

Also read issue 259 for other possible changes for version 2.0.0:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.