REMOVE support for apt-sources-list
This package already inherits from basic faces.  There seems to be no
need to extend support for it at the theme level, also because of its
niche utility (I mean, how often do you edit your apt sources and how
difficult is it to just live with the standard font-lock faces for a few
minutes?).

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.