REMOVE support for easy-kill
This package inherits from base faces and should be okay without
explicit theme support.

I am actually not sure we should remove this, but I am willing to take
the risk.  If someone objects to this change, they are free to contact
me and I shall reconsider.

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.