REMOVE support for rainbow-identifiers
This package has not been updated in seven years and has no other
activity in its git repo.  This may be because it is stable and "just
works".  Or it has been de facto deprecated.

Though I cannot be certain, I am willing to take the risk and remove it
from the themes.  If someone objects to this change, they are welcome to
inform me about and I shall reconsider.

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.