REMOVE support for highlight-blocks
This package does not even have a README.  It was last updated almost
three years ago and there seems to be no activity in its git repo.

I am thus willing to take the risk of removing it.  If anyone objects to
this change, they are welcome to inform me about it and I shall
reconsider.

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.