REMOVE support for highlight-tail
This package is hosted on EmacsWiki and I cannot know whether there is
any activity around its development.  Its history of changes stops at
2007, while its Commentary section makes references to speed issues.

While I am not sure about its status, I am thus willing to take the risk
of removing it.  If anyone objects to this change, they are welcome to
inform me about it and I shall reconsider.

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.