REMOVE support for highlight-symbol
This package is no longer maintained.

Also, it may be easier to just use the built-in highlight-regexp
functionality (bound to M-s h r and can also be extended from
isearch---contact me if you need help).  Read the manual by evaluating:

    (info "(emacs) Highlight Interactively")

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.