REMOVE support for diredc
This is a massive package that seems like a de facto fork of Dired.  I
cannot promise support for it in its current stage, so it is better to
remove it from the themes until someone who uses it in earnest comes
along and helps cover its use-case.

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.