REMOVE support for phi-search
Its homepage points to a website that seems to have nothing to do with
the package.  What is going on?

I am not sure about this decision, but am willing to take the risk of
removing it.  If anyone objects to this change, they are welcome to
inform me about it and I shall reconsider.

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.