REMOVE support for isl (isearch-light)
I remember adding support for this[1] on the premise that a package
would eventually become available on some ELPA.  This has not happened
and I am not sure what to expect.

[1] <https://github.com/thierryvolpiatto/isearch-light>

As such, removing it may be a drastic measure, though I am willing to
take the risk.  If anyone objects to this change, they are welcome to
inform me about it and I shall reconsider.

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.