REMOVE support for objed
This package is subject to a rewrite.  Its three faces inherit from
basic faces that we support.  While we would have reasons to style
things differently, I think that we are better off removing it from the
themes until further notice.

Quote from its README.asc:

    Development is currently paused but I have plans for a complete overhaul
    of the project. Objed is completely usable in its current state but note
    that its public interface is subject to change.

    <https://github.com/clemera/objed/blob/master/README.asc>

We should be more selective with our choices, as noted in issue 259:
<https://gitlab.com/protesilaos/modus-themes/-/issues/259>.