BREAKING remove small scale from Org agenda
It never worked well, as it caused alignment issues.  There is nothing
we can do at the theme level to fix the issue: Org should have another
way to aligning to a specific length instead of font-based spacing.

This is part of the progress towards version 2.0.0 of the themes.  See
issue 259: <https://gitlab.com/protesilaos/modus-themes/-/issues/259>.