Stop treating Latex sections as headings
This is because unlike Org/Outline/Markdown Latex is basically source
code, so the sectioning does not work the same way it does for those
lightweight markup/outlining modes.

Furthermore, font-latex.el defines 'font-latex-fontify-sectioning' which
can be used to control the scale of those headings.  It makes sense for
the themes to not interfere with that design and just allow users to
customise things uniformly regardless of the active theme.

Thanks to Gustavo Barros for the detailed feedback in issue 265:
<https://gitlab.com/protesilaos/modus-themes/-/issues/265>.