Change default style of Org block (no backgrounds)
As I explained in Emacs bug#52587,[1] Org has code that overrides themes
which prefer not to extend the block delimiter faces to the edge of the
window.  This practically means that we cannot have backgrounds for
those lines and keep them limited to the stretch of area covered by
their text.

As such, the default for Org block delimiter lines now is a gray
foreground with no distinct background colour.  The user option
'modus-themes-org-blocks' provides "blocky" alternatives that use
background colours---those extend to the edge of the window.

[1] <https://debbugs.gnu.org/cgi/bugreport.cgi?bug=52587#46>