Make Eldoc current argument more noticeable
The blue can be hard to tell apart in some cases, especially because it
is a common colour that is used elsewhere in the themes.  Whereas this
shade of red is easier to discern.  Hopefully it does not feel too
alarmist (in which case we need to revise it).

Thanks to Rudolf Adamkovič for the feedback in issue 275:
<https://gitlab.com/protesilaos/modus-themes/-/issues/275>.