Make echo area current arg more prominent
In commit be886d3 I wrote:

    The red colour feels too aggressive.  I tried it for two full days and
    still cannot get used to it.

    This revises commits aa62a2a, fe911a4.  It relates to issue 275:
    <https://gitlab.com/protesilaos/modus-themes/-/issues/275>.

Though the latest feedback I got from Rudolf Adamkovič includes
information I had not considered before: that one may check the
indicator using their peripheral vision.  Whereas I have the habit of
actually focusing on the echo area.

Couched in those terms, the red hue is an appropriate choice: it may
still be too prominent for my use-case, but even so it remains usable.
We thus have one case where the red is essential and another where it is
not impeding the given functionality.

If someone thinks the red is intolerable, then we will have to introduce
a defcustom.

At any rate, I don't want to have a background there by default because
it is easy to mistake it for a mouse highlight or something that has to
do with minibuffer completion.  An underline is problematic as well: it
is not visible in fullscreen view and it does not work as well when
using M-: (the eval-expression command) where the Eldoc indicators move
to the mode line.