Stop making keys look like boxes
We revert to the old style we were using before the introduction of the
'help-key-binding' face (Emacs 28).

* * *

By default Emacs 28 or higher will render all key bindings it identifies
with a box around them.  The idea is to make them look like keys, which
I never really liked because without padding you get a very tight space
between the character and the box's borders which can look weird at
small point sizes.

I tried following the default style for a few months and have concluded
that it is not good enough for our purposes (my preferences
notwithstanding):

* The box attribute does not work in terminal emulators.  This means
  that keys only get a subtle grey background and the default
  foreground, which can be hard to make them stand out from their
  surrounding text if the font height is small and/or the keybinding is
  short (e.g. a single character).

* The box and grey background combination limits our options when we
  need to colour-code different types of keys.  For example, the
  which-key package can show TAB as T and applies to it a different face
  to make the distinction obvious.  In that case, the presence of the
  tight box makes the use of a bold weight inappropriate: the character
  and the box's borders seem to overlap.  While the grey background
  limits our choice of colour as, for instance, yellow never looks good
  against it.  Same principle for interfaces that can have colour-coded
  keys like transient and hydra, where we lose much-needed flexibility.