Tweak modus-vivendi bg-paren-match colour
I wanted to increase its distance relative to the main background, just
to be sure that it is easier to spot.  This is achieved by moving the
hueness from the yellow to the magenta side of the spectrum.

Overall, the change is subtle and has no major impact on the contrast
ratio relative to the main background and foreground (we need to
consider both due to the specifics of show-paren-mode (and related)).

The results:

    |         | #000000 | #ffffff | #000000 | #ffffff |
    |---------+---------+---------+---------+---------|
    | #5f362f |    2.06 |   10.22 |   37904 |  333060 |
    | #6f3355 |    2.28 |    9.21 |   58282 |  291037 |

The TBLFM formula for this table (org-mode notation):

    $2='(Λ $1 @1$2);%.2f :: $3='(Λ $1 @1$3);%.2f :: $4='(Δ $1 @1$4) :: $5='(Δ $1 @1$5)

The Greek letters mean:

    (defalias 'Λ #'modus-themes-contrast)
    (defalias 'Δ #'color-distance)