REMOVE translation layer for all old variables
The original plan was to remove those during the transition to version
2.0.0 (about a month ago) though I changed my mind thinking they would
not pose a longer-term problem.

New information by Mark Bestley in issue 272[1] shows that this kind of
complexity can lead to errors, so it is better to keep things simple and
ask users to configure all user options based on the up-to-date
documentation.  The manual includes a snippet with all sample code.

Evaluate this if you have it installed:

    (info "(modus-themes) Customization Options")

Otherwise read it online:
<https://protesilaos.com/emacs/modus-themes#h:bf1c82f2-46c7-4eb2-ad00-dd11fdd8b53f>.

Link to Mark Bestley's feedback:
<https://gitlab.com/protesilaos/modus-themes/-/issues/272#note_826725412>.
Also thanks to Saša Janiška for the feedback in that issue.