Make the font slant unopinionated for consistency
This does not hardcode ':slant italic' to the 'italic' face, in the
same way we do not hardcode ':weight bold' to the 'bold' face.

Such a design allows users to configure those faces and have the desired
slant/weight (and even font family) apply consistently throughout the
theme.

Read the manual for further details:
<https://protesilaos.com/emacs/modus-themes#h:2793a224-2109-4f61-a106-721c57c01375>.

Thanks to user derek-upham for pointing out the inconsistency in issue
21 at the GitHub mirror: <https://github.com/protesilaos/modus-themes/issues/21>.