Refine compilation-mode colours and fallback face
Before the fallback face was the generic 'underline'.  Now we use the
'link' face which is the correct one for this case.

The colours are revised to work better in all the relevant contexts,
while we no longer try to colour-code line and column numbers (it is not
effective).

Thanks to Rudolf Adamkovič for the feedback in issue 277:
<https://gitlab.com/protesilaos/modus-themes/-/issues/277>.