Stop being opinionated about compilation buffers
This effectively annuls commit 718653f: we should keep the default where
Emacs applies the generic 'underline' face, otherwise we get awkward
colour combinations.

In the manual we explain how the user can disable those underlines: it
is a user option and we ought to respect it.

Read issue 277 for the ongoing discussion:
<https://gitlab.com/protesilaos/modus-themes/-/issues/277>.