Add colour back to mode-line-emphasis
After some further testing, I feel that the added colour helps with the
emphasis.  The reason is that we normally don't have our eyes fixed on
the mode line, so bold alone may not be enough to call attention to
itself.

This effectively reverts commit 1b110f5, though it changes the
foreground from a shade of blue to magenta.