REFACTOR modus-themes-completions to an alist type
The user option 'modus-themes-completions' now accepts an alist value
instead of a symbol.  For example:

    (setq modus-themes-completions '((matches . (extrabold))
                                     (selection . (semibold accented))
                                     (popup . (accented intense))))

The doc string or the manual explains the details.

All old values, such as 'opinionated' and 'moderate' ARE IGNORED.

This change is necessary as the old design was inflexible and led to
awkward compromises with the choice of colours and overall looks.

This is a MAJOR CHANGE for Ivy and Helm, as their default looks are now
subtle, aligning them with all other completions UIs.  Users who
preferred the old style can use something like this:

    (setq modus-themes-completions '((matches . (background intense))
                                     (selection . (accented intense))
                                     (popup . (accented intense))))

Or simply:

    (setq modus-themes-completions '((t background intense accented)))

Again, please read the docs.

Furthermore, the new 'modus-themes-completions' finally harmonises all
completion UIs supported by the themes.  This means that it also applies
to Corfu, Company, and the Popup library.

The faces of all affected packages have been reviewed to ensure
theme-wide consistency and/or keep things simple.

Thanks to Daniel Mendler for the feedback in issue 278:
<https://gitlab.com/protesilaos/modus-themes/-/issues/278>.