Refine intense+background matches in completions
This pertains to the new design of 'modus-themes-completions', which
accepts an alist value.  Please read its doc string for all the details,
including sample code.  Also check the release notes for version 2.2.0
which were published earlier today:
<https://protesilaos.com/codelog/2022-02-23-modus-themes-2-2-0/>.

The 'matches' key accepts a list of properties, among which are
'intense' and 'background'.  When those two are combined, they now
produce a more colourful result than before.  This follows a natural
progression from the style of 'background' WITHOUT 'intense'.

Thanks to Ryan (ryanask) for the feedback in issue 278:
<https://gitlab.com/protesilaos/modus-themes/-/issues/278>.