Add text-also property to modus-themes-completions
This adds the option for an extra colour to the selected line's text.
It basically is yet another way of amplifying its styling.

I used this opportunity to deprecate the 'modus-themes--completions' and
replace it with bespoke functions for the relevant keys in
'modus-themes-completions'.

Thanks to Morgan Willcock whose feedback in issue 278 inspired me to add
the 'text-also' property:
<https://gitlab.com/protesilaos/modus-themes/-/issues/278>.